requirements.txt
Django==4.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers